Modifies journalSetsByYear
The issuesInYear function now returns only 52 issues and ensures that
they have publication dates in the specified year. In other words,
issues published at the end of the prior year are no longer included.
The journalSetsByYear function now returns 26 sets that cover all
issues in the specified year. This includes a final set of possible
'stragler' issues that may or may not be published at all.