Fixes bug in reading end of titles
Some titles were not matching correctly, because the journal name was
not being recognized. This is because the journal names can get split
over multiple lines. So, the newlines need to be removed and replaced
with spaces in a way that does not introduce double spaces. This was
accomplished by using (unwords . words).