Fixes online page numbers
Previously online page numbers used a default string. To ensure the
page numbers are all distinct, the doi suffix is used instead along
with "1" for the first page number and "0" for the last. The
difference is negative, which can be used to easily identify it as
an online-only article.

The prefix and digit suffix are separated by a ':' in text
representations unless the prefix is empty, in which case the suffix
is the just the digits.