Start dealing with new PubMed website
On May 18, 2020, a new version of PubMed.gov was released. This
version did not return searches using plain text format, which the
previous version used to obtain tables of contents. So, a new
approach was required to obtain tables of contens from PubMed using
the E-Utilities API, which operates very differently and returns
search and summary results using JSON. This commit starts the process
of dealing with this by providing the functions that correctly
construct the necessary API calls for obtaining the PMIDs for
citations in a given issue and then obtaining document summaries
from these PMIDs.