Improves parsing of HTML in PubMed JSON strings
PubMed replaces html elements such as '<' and '>' with their
corresponding html entities. I don't know why, but this may be done
because the default data format is XML rather than JSON, in which
case the use of entities would be required. In any case, we need to
parse these entities out as the correct characters for them to render
properly. This commit adjusts the json parser to do this.

However, the new parser relies on PubMed ensuring that the all of
their html tags are correctly formatted (e.g., every '<' has a
closing '>'). The new parser does not do any of this checking.
Furthemore, I have not encountered any title that has a legitimate
'<' or '>' symbol. However, I assume it would be encoded as
"&amp;lt;". In this case, the parser will parse this as "&lt;",
which is what we want.