Fixes bug preventing parsing of publication dates
In the absence of a publication date field (pubdate) in the ESummary
json documents, the epubdate field was being used to set the date of
a citation. However, sometimes the pubdate field is present but does
not provide a complete year-month-day date. This would fail to parse
without ever checking the epubdate field and the entire citation
would then fail to parse. This commit fixes this by first checking
whether the pubdate field is present and if so trying to parse it to
a valid date. If this fails, then the epubdate field is checked.