Changes doi formatting in Citations
Previously the 'https://www.doi.org/' prefix was added to doi values
in the doi Citation field when parsing PubMed json. However, this
prefix was not being added to citations obtained by direct doi lookup
and parsing the RIS. This bug resulted in the citations not being
found in the corresponding html documents.

Rather than adding the prefix during parsing, which would entail
adding the prefix at both PubMed json parsing as well as at RIS
parsing, it was removed from the parsing step so just the base doi is
stored in the Citation. The prefix is then added at the time the
citation is formatted as html or Markdown.

Note that the prefix was also changed from 'https://www.doi.org/' to
'https://doi.org/', which is preferred.