Modifies RIS parser to handle nonstandard formats
The RIS format is *supposed* to have every key-value pair on a single
line; however, some doi lookups return RIS where a value is broken
over multiple lines with leading spaces on each line. To handle this,
failure to parse a RIS key is assumed to represent a line break in a
field; however, it must still end with a new line. After parsing the
RIS, the line breaks are reassembled and the leading white space is
removed before converting it to a citation.