Fixes RIS parsing when there is no journal
Some publications (e.g., those at bioRxiv) do not have associated
journals. This was causing the RIS reader to fail. This commit fixes
this by first checking for a journal and if absent using the
publisher annotated with the RIS type as the journal title. If there
is no publisher, then "No Publisher" is used with the RIS type. The
RIS type is always present in valid RIS.